refactor: extract the Web API layer into src/api/
main.rs 3599 -> 2759. The HTTP layer was ~845 lines sitting in the middle of
the binary, between the input handlers and the terminal setup, which meant
anyone reading the app loop had to scroll past every paging helper to get
there.

It moves out as-is: one module per thing fetched, so the file to open is the
one named after it.

  mod.rs      transport core - client, token, get_json, get_json_cached, cover
  library.rs  home, recent, playlists, albums, artists, liked + paging
  search.rs   search
  detail.rs   drill-in: artist top tracks, artist albums, album, playlist
  queue.rs    the live play queue
  track.rs    one track's metadata
  lyrics.rs   lrclib
  playback.rs server-side playback state, transfer, restore

Same one-way dependency as src/ui/, pointing the other way: everything in
here talks HTTP and hands plain data back over channels, and nothing in here
touches App or the render tree.

Pure move. No logic changed - the only edits are pub(crate) visibility, the
use statements, and rustfmt rewrapping three signatures that pub(crate)
pushed past 100 columns. Section banner comments became module docs.

139 tests still pass; fmt, clippy and test all clean with --all-features
under -D warnings.
