+++ src/main.rs
@@ -5,6 +5,10 @@
 //! Multi-section library (playlists / liked / albums / artists), shuffle, repeat,
 //! and a live queue view.
 
+/// The Spotify Web API layer. Talks HTTP, hands plain data back over channels.
+/// Lives in the binary (not the library) because it speaks the model types
+/// defined here.
+mod api;
 /// The render tree. Reads `App`, writes `FrameOut`; never the other way round.
 /// Lives in the binary (not the library) because it needs `App`, which is here.
 mod ui;
@@ -27,6 +31,7 @@ use ratatui::widgets::{Block, Clear, Paragraph};
 use ratatui::Frame;
 use ratatui_image::picker::Picker;
 
+use api::*;
 use myx::anim::ThemeFade;
 use myx::audio::NUM_BANDS;
 use myx::components::{gradient_line, gradient_progress, left_bar_block};
@@ -2722,851 +2727,6 @@ fn apply_meta(
                 .map(|n| Duration::from_millis(n.duration_ms as u64)),
         });
     }
-}
-
-// ------------------------------------------------------------------ web api
-
-fn token_of(webapi: &Arc<Mutex<WebApi>>) -> Option<String> {
-    // Refresh when the token is expiring so long sessions don't silently go
-    // read-only (audit H1). Only holds the lock across the network call in the
-    // rare refresh window; otherwise this is just a cheap clone.
-    let token = {
-        let mut w = webapi.lock().ok()?;
-        match w.valid_token() {
-            Ok(t) => t,
-            Err(_) => w.cached_token(),
-        }
-    };
-    (!token.is_empty()).then_some(token)
-}
-
-/// Base URL every Spotify Web API call is built from.
-const API: &str = "https://api.spotify.com/v1";
-
-/// How long to wait before retrying a 429, or `None` to give up now.
-///
-/// Spotify hands development-mode apps hour-long `Retry-After` values once a
-/// quota is spent, and sleeping on those froze a drill-in for minutes before
-/// failing anyway. Only brief backoffs are worth waiting out.
-fn retry_delay(retry_after: Option<u64>) -> Option<Duration> {
-    match retry_after.unwrap_or(3) {
-        secs if secs <= 5 => Some(Duration::from_secs(secs + 1)),
-        _ => None,
-    }
-}
-
-/// GET a JSON endpoint, retrying on 429 (respecting Retry-After).
-fn get_json(
-    client: &reqwest::blocking::Client,
-    url: &str,
-    token: &str,
-) -> Option<serde_json::Value> {
-    for _ in 0..5 {
-        let resp = match client.get(url).bearer_auth(token).send() {
-            Ok(r) => r,
-            Err(e) => {
-                liblog(format!("api: {url} transport error: {e}"));
-                return None;
-            }
-        };
-        if resp.status().as_u16() == 429 {
-            let after = resp
-                .headers()
-                .get("retry-after")
-                .and_then(|v| v.to_str().ok())
-                .and_then(|s| s.parse::<u64>().ok());
-            let Some(wait) = retry_delay(after) else {
-                liblog(format!(
-                    "api: {url} -> 429, retry-after {after:?}s, giving up"
-                ));
-                return None;
-            };
-            std::thread::sleep(wait);
-            continue;
-        }
-        if !resp.status().is_success() {
-            // Swallowing the status made a dead endpoint look like an empty one.
-            liblog(format!("api: {url} -> HTTP {}", resp.status().as_u16()));
-            return None;
-        }
-        return resp.json::<serde_json::Value>().ok();
-    }
-    None
-}
-
-/// How long a cached catalogue response counts as fresh. Discographies and
-/// track lists change on the order of weeks; a day keeps browsing off the
-/// network without anyone noticing the lag.
-const CATALOGUE_TTL: Duration = Duration::from_secs(24 * 60 * 60);
-
-/// `get_json` for catalogue reads, served from disk when possible.
-///
-/// On a failed request an expired entry is used rather than nothing — that is
-/// the case that matters, since a spent quota is exactly when the network stops
-/// answering and the artist page would otherwise come up empty.
-fn get_json_cached(
-    client: &reqwest::blocking::Client,
-    url: &str,
-    token: &str,
-) -> Option<serde_json::Value> {
-    if let Some(body) = myx::httpcache::get(url, Some(CATALOGUE_TTL)) {
-        return serde_json::from_str(&body).ok();
-    }
-    match get_json(client, url, token) {
-        Some(v) => {
-            myx::httpcache::put(url, &v.to_string());
-            Some(v)
-        }
-        None => {
-            let stale = myx::httpcache::get(url, None)?;
-            liblog(format!("api: {url} failed; serving cached copy"));
-            serde_json::from_str(&stale).ok()
-        }
-    }
-}
-
-/// Album art bytes, from disk when they've been seen before.
-fn fetch_cover(client: &reqwest::blocking::Client, url: &str) -> Option<Vec<u8>> {
-    if let Some(bytes) = myx::httpcache::get_bytes(url) {
-        return Some(bytes);
-    }
-    let resp = client.get(url).send().ok()?;
-    // An error page cached as image bytes would be a permanently broken cover:
-    // the entry never expires, because the URL only changes with the picture.
-    if !resp.status().is_success() {
-        liblog(format!("cover: {url} -> HTTP {}", resp.status().as_u16()));
-        return None;
-    }
-    let bytes = resp.bytes().ok()?.to_vec();
-    myx::httpcache::put_bytes(url, &bytes);
-    Some(bytes)
-}
-
-/// Fetch the library incrementally: fast sections first, Liked streamed in
-/// chunks so the UI is usable within ~1s instead of waiting for everything.
-fn spawn_library_fetch(
-    webapi: Arc<Mutex<WebApi>>,
-    tx: flume::Sender<(Section, Vec<LibItem>)>,
-    done_tx: flume::Sender<bool>,
-) {
-    // Clone the already-refreshed access token BEFORE spawning. This removes the
-    // shared WebApi mutex from the worker entirely (the stuck-library root cause).
-    let token_opt = token_of(&webapi);
-    liblog(format!(
-        "spawn_library_fetch: token={}",
-        token_opt.as_ref().map_or("missing", |_| "ok")
-    ));
-    std::thread::Builder::new()
-        .name("myx-library".to_string())
-        .spawn(move || {
-            liblog("worker: entered");
-            let Some(token) = token_opt else {
-                liblog("worker: no token; aborting");
-                let _ = done_tx.send(false);
-                return;
-            };
-
-            let client = reqwest::blocking::Client::builder()
-                .timeout(Duration::from_secs(12))
-                .build()
-                .expect("build library HTTP client");
-            let mut got_any = false;
-            let track_from = |t: &serde_json::Value| -> Option<LibItem> {
-                Some(LibItem::track(
-                    t["name"].as_str()?.to_string(),
-                    t["artists"][0]["name"].as_str().unwrap_or("").to_string(),
-                    t["uri"].as_str()?.to_string(),
-                ))
-            };
-            let artist_from = |a: &serde_json::Value| -> Option<LibItem> {
-                Some(LibItem::ctx(
-                    a["name"].as_str()?.to_string(),
-                    String::new(),
-                    a["uri"].as_str()?.to_string(),
-                ))
-            };
-            let album_from = |a: &serde_json::Value| -> Option<LibItem> {
-                Some(LibItem::ctx(
-                    a["name"].as_str()?.to_string(),
-                    format!("album · {}", a["artists"][0]["name"].as_str().unwrap_or("")),
-                    a["uri"].as_str()?.to_string(),
-                ))
-            };
-
-            // Home: a curated mix — recently played, top tracks, top artists, new releases.
-            liblog("worker: fetching home/recent");
-            let mut home: Vec<LibItem> = Vec::new();
-            let recent5 = fetch_all_pages(
-                &client,
-                &format!("{API}/me/player/recently-played?limit=10"),
-                &token,
-                None,
-                1,
-                |it| track_from(&it["track"]),
-            );
-            if !recent5.is_empty() {
-                home.push(LibItem::header("Recently Played"));
-                home.extend(recent5.into_iter().take(6));
-            }
-            let top_tracks = fetch_all_pages(
-                &client,
-                &format!("{API}/me/top/tracks?limit=10"),
-                &token,
-                None,
-                1,
-                |t| track_from(t),
-            );
-            if !top_tracks.is_empty() {
-                home.push(LibItem::header("Your Top Tracks"));
-                home.extend(top_tracks.into_iter().take(8));
-            }
-            let top_artists = fetch_all_pages(
-                &client,
-                &format!("{API}/me/top/artists?limit=10"),
-                &token,
-                None,
-                1,
-                |a| artist_from(a),
-            );
-            if !top_artists.is_empty() {
-                home.push(LibItem::header("Your Top Artists"));
-                home.extend(top_artists.into_iter().take(6));
-            }
-            // `/browse/new-releases` was removed in February 2026.
-            got_any |= !home.is_empty();
-            liblog(format!("worker: home done, {} rows", home.len()));
-            let _ = tx.send((Section::Home, home));
-
-            let recent = fetch_all_pages(
-                &client,
-                &format!("{API}/me/player/recently-played?limit=50"),
-                &token,
-                None,
-                1,
-                |it| track_from(&it["track"]),
-            );
-            got_any |= !recent.is_empty();
-            let _ = tx.send((Section::Recent, recent));
-
-            let playlists = fetch_all_pages(
-                &client,
-                &format!("{API}/me/playlists?limit=50"),
-                &token,
-                None,
-                10,
-                |it| {
-                    Some(LibItem::ctx(
-                        it["name"].as_str()?.to_string(),
-                        playlist_subtitle(
-                            it["owner"]["display_name"].as_str().unwrap_or(""),
-                            playlist_total(it),
-                        ),
-                        it["uri"].as_str()?.to_string(),
-                    ))
-                },
-            );
-            got_any |= !playlists.is_empty();
-            let _ = tx.send((Section::Playlists, playlists));
-
-            let albums = fetch_all_pages(
-                &client,
-                &format!("{API}/me/albums?limit=50"),
-                &token,
-                None,
-                10,
-                |it| album_from(&it["album"]),
-            );
-            got_any |= !albums.is_empty();
-            let _ = tx.send((Section::Albums, albums));
-
-            let artists = fetch_all_pages(
-                &client,
-                &format!("{API}/me/following?type=artist&limit=50"),
-                &token,
-                Some("artists"),
-                5,
-                |it| artist_from(it),
-            );
-            got_any |= !artists.is_empty();
-            let _ = tx.send((Section::Artists, artists));
-
-            // Liked can be huge — stream it in as pages arrive so the count climbs live.
-            // Prepend Shuffle/Play action rows (shuffle first).
-            let mut liked: Vec<LibItem> = vec![
-                LibItem::play("▶︎  Play Liked Songs".into(), "myx:action:liked-play".into()),
-                LibItem::header("Songs"),
-            ];
-            let mut url = Some(format!("{API}/me/tracks?limit=50"));
-            let mut pages = 0;
-            while let Some(u) = url.take() {
-                if pages >= 100 {
-                    break;
-                }
-                let Some(v) = get_json(&client, &u, &token) else {
-                    break;
-                };
-                for it in v["items"].as_array().into_iter().flatten() {
-                    if let Some(li) = track_from(&it["track"]) {
-                        liked.push(li);
-                    }
-                }
-                url = v["next"].as_str().map(String::from);
-                pages += 1;
-                if pages % 3 == 0 {
-                    let _ = tx.send((Section::Liked, liked.clone()));
-                }
-            }
-            got_any |= liked.len() > 2; // beyond the two action rows
-            let _ = tx.send((Section::Liked, liked));
-
-            liblog(format!("worker: all done got_any={got_any}"));
-            let _ = done_tx.send(got_any);
-        })
-        .expect("spawn library worker");
-}
-
-/// One `/playlists/{id}/items` entry -> `LibItem`.
-///
-/// The payload nests the track under `item`; the older `/tracks` endpoint used
-/// `track`. Both are accepted so this keeps working whichever shape is served.
-///
-/// `None` skips the row (`fetch_all_pages` filters rather than aborting), which
-/// is what we want for entries with no playable track: `null` for items removed
-/// from the catalogue, and region-locked or malformed rows.
-fn parse_playlist_track(it: &serde_json::Value) -> Option<LibItem> {
-    let t = if it["item"].is_object() {
-        &it["item"]
-    } else {
-        &it["track"]
-    };
-    Some(LibItem::track(
-        t["name"].as_str()?.to_string(),
-        t["artists"][0]["name"].as_str().unwrap_or("").to_string(),
-        t["uri"].as_str()?.to_string(),
-    ))
-}
-
-/// Track count from a playlist object. Spotify renamed the field `tracks` ->
-/// `items` alongside the `/tracks` -> `/items` endpoint move; read the new name
-/// first and fall back so both shapes work.
-fn playlist_total(p: &serde_json::Value) -> Option<u64> {
-    p["items"]["total"]
-        .as_u64()
-        .or_else(|| p["tracks"]["total"].as_u64())
-}
-
-/// Playlist row subtitle: `"142 · owner"`, or just the owner when the API omits
-/// the count.
-///
-/// Count first, deliberately: the row renderer truncates the subtitle tail-first
-/// in a narrow pane, and the count is both short and the more informative half —
-/// the owner is frequently the same name on every row.
-fn playlist_subtitle(owner: &str, total: Option<u64>) -> String {
-    match total {
-        Some(n) if owner.is_empty() => n.to_string(),
-        Some(n) => format!("{n} · {owner}"),
-        None => owner.to_string(),
-    }
-}
-
-fn fetch_all_pages(
-    client: &reqwest::blocking::Client,
-    first_url: &str,
-    token: &str,
-    nested: Option<&str>,
-    max_pages: usize,
-    parse: impl Fn(&serde_json::Value) -> Option<LibItem>,
-) -> Vec<LibItem> {
-    let mut out = Vec::new();
-    let mut url = Some(first_url.to_string());
-    let mut pages = 0;
-    while let Some(u) = url.take() {
-        if pages >= max_pages {
-            break;
-        }
-        let Some(v) = get_json(client, &u, token) else {
-            break;
-        };
-        let node = match nested {
-            Some(k) => &v[k],
-            None => &v,
-        };
-        for it in node["items"].as_array().into_iter().flatten() {
-            if let Some(li) = parse(it) {
-                out.push(li);
-            }
-        }
-        url = node["next"].as_str().map(String::from);
-        pages += 1;
-    }
-    out
-}
-
-fn spawn_queue_fetch(webapi: Arc<Mutex<WebApi>>, tx: flume::Sender<Vec<(String, String)>>) {
-    tokio::task::spawn_blocking(move || {
-        let q = match token_of(&webapi) {
-            Some(token) => fetch_queue_blocking(&token),
-            None => Vec::new(),
-        };
-        let _ = tx.send(q);
-    });
-}
-
-/// Returns (display, uri) for each queued item.
-fn fetch_queue_blocking(token: &str) -> Vec<(String, String)> {
-    let client = http_client();
-    let Some(v) = get_json(&client, &format!("{API}/me/player/queue"), token) else {
-        return Vec::new();
-    };
-    v["queue"]
-        .as_array()
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|it| {
-                    let name = it["name"].as_str()?;
-                    let uri = it["uri"].as_str()?.to_string();
-                    let artist = it["artists"][0]["name"].as_str().unwrap_or("");
-                    Some((format!("{name} — {artist}"), uri))
-                })
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn fetch_track_meta(webapi: &Arc<Mutex<WebApi>>, track_id: &str) -> TrackMeta {
-    let uri = format!("spotify:track:{track_id}");
-    let empty = || TrackMeta {
-        uri: uri.clone(),
-        title: String::new(),
-        artist: String::new(),
-        album: String::new(),
-        duration_ms: 0,
-        image: TrackImage {
-            image: None,
-            url: None,
-        },
-        theme: None,
-    };
-    let Some(token) = token_of(webapi) else {
-        return empty();
-    };
-    let client = http_client();
-    let Some(v) = get_json_cached(&client, &format!("{API}/tracks/{track_id}"), &token) else {
-        return empty();
-    };
-
-    let title = v["name"].as_str().unwrap_or("").to_string();
-    let artist = v["artists"][0]["name"].as_str().unwrap_or("").to_string();
-    let album = v["album"]["name"].as_str().unwrap_or("").to_string();
-    let duration_ms = v["duration_ms"].as_u64().unwrap_or(0) as u32;
-    let cover_url = v["album"]["images"][0]["url"].as_str().map(String::from);
-
-    let image = cover_url
-        .clone()
-        .and_then(|u| fetch_cover(&client, &u))
-        .and_then(|bytes| image::load_from_memory(&bytes).ok());
-    let theme = image.as_ref().map(|img| derive_theme(img, "album ✦"));
-
-    TrackMeta {
-        uri,
-        title,
-        artist,
-        album,
-        duration_ms,
-        image: TrackImage {
-            image,
-            url: cover_url,
-        },
-        theme,
-    }
-}
-
-// --- Search ---
-
-fn spawn_search(webapi: Arc<Mutex<WebApi>>, query: String, tx: flume::Sender<Vec<LibItem>>) {
-    tokio::task::spawn_blocking(move || {
-        let results = match token_of(&webapi) {
-            Some(token) => search_blocking(&token, &query),
-            None => Vec::new(),
-        };
-        let _ = tx.send(results);
-    });
-}
-
-fn search_blocking(token: &str, query: &str) -> Vec<LibItem> {
-    let client = http_client();
-    let url = format!(
-        "{API}/search?q={}&type=track,artist,album,playlist&limit=6",
-        urlencode(query)
-    );
-    let Some(v) = get_json(&client, &url, token) else {
-        return Vec::new();
-    };
-
-    let mut out = Vec::new();
-
-    let songs: Vec<LibItem> = v["tracks"]["items"]
-        .as_array()
-        .into_iter()
-        .flatten()
-        .filter_map(|t| {
-            Some(LibItem::track(
-                t["name"].as_str()?.to_string(),
-                t["artists"][0]["name"].as_str().unwrap_or("").to_string(),
-                t["uri"].as_str()?.to_string(),
-            ))
-        })
-        .collect();
-    let artists: Vec<LibItem> = v["artists"]["items"]
-        .as_array()
-        .into_iter()
-        .flatten()
-        .filter_map(|a| {
-            Some(LibItem::ctx(
-                a["name"].as_str()?.to_string(),
-                String::new(),
-                a["uri"].as_str()?.to_string(),
-            ))
-        })
-        .collect();
-    let albums: Vec<LibItem> = v["albums"]["items"]
-        .as_array()
-        .into_iter()
-        .flatten()
-        .filter_map(|al| {
-            Some(LibItem::ctx(
-                al["name"].as_str()?.to_string(),
-                al["artists"][0]["name"].as_str().unwrap_or("").to_string(),
-                al["uri"].as_str()?.to_string(),
-            ))
-        })
-        .collect();
-    let playlists: Vec<LibItem> = v["playlists"]["items"]
-        .as_array()
-        .into_iter()
-        .flatten()
-        .filter_map(|p| {
-            Some(LibItem::ctx(
-                p["name"].as_str()?.to_string(),
-                playlist_subtitle(
-                    p["owner"]["display_name"].as_str().unwrap_or(""),
-                    playlist_total(p),
-                ),
-                p["uri"].as_str()?.to_string(),
-            ))
-        })
-        .collect();
-
-    for (title, group) in [
-        ("Songs", songs),
-        ("Artists", artists),
-        ("Albums", albums),
-        ("Playlists", playlists),
-    ] {
-        if !group.is_empty() {
-            out.push(LibItem::header(title));
-            out.extend(group);
-        }
-    }
-    out
-}
-
-// --- Lyrics (lrclib) ---
-
-fn fetch_lyrics_blocking(
-    artist: &str,
-    title: &str,
-    album: &str,
-    duration_ms: u32,
-) -> (Vec<(u32, String)>, bool) {
-    let client = http_client();
-    let url = format!(
-        "https://lrclib.net/api/get?artist_name={}&track_name={}&album_name={}&duration={}",
-        urlencode(artist),
-        urlencode(title),
-        urlencode(album),
-        duration_ms / 1000
-    );
-    let Ok(resp) = client
-        .get(&url)
-        .header("User-Agent", "myx (terminal spotify player)")
-        .send()
-    else {
-        return (Vec::new(), false);
-    };
-    if !resp.status().is_success() {
-        return (Vec::new(), false);
-    }
-    let Ok(v) = resp.json::<serde_json::Value>() else {
-        return (Vec::new(), false);
-    };
-
-    if let Some(synced) = v["syncedLyrics"].as_str().filter(|s| !s.is_empty()) {
-        return (parse_lrc(synced), true);
-    }
-    if let Some(plain) = v["plainLyrics"].as_str().filter(|s| !s.is_empty()) {
-        let lines = plain.lines().map(|l| (0u32, l.to_string())).collect();
-        return (lines, false);
-    }
-    (Vec::new(), false)
-}
-
-// --- Drill-in detail (artist / album / playlist) ---
-
-fn spawn_detail_fetch(
-    webapi: Arc<Mutex<WebApi>>,
-    uri: String,
-    name: String,
-    tx: flume::Sender<(String, String, Vec<LibItem>)>,
-) {
-    tokio::task::spawn_blocking(move || {
-        if let Some(token) = token_of(&webapi) {
-            let (title, items) = fetch_detail_blocking(&token, &uri, &name);
-            let _ = tx.send((uri, title, items));
-        }
-    });
-}
-
-/// An artist's most popular tracks. `/artists/{id}/top-tracks` was removed in
-/// February 2026, so this searches instead — also popularity-ranked, ten max.
-fn artist_top_tracks(
-    client: &reqwest::blocking::Client,
-    token: &str,
-    id: &str,
-    name: &str,
-) -> Vec<LibItem> {
-    let url = format!(
-        "{API}/search?q={}&type=track&limit=10",
-        urlencode(&format!("artist:{name}"))
-    );
-    let Some(v) = get_json_cached(client, &url, token) else {
-        return Vec::new();
-    };
-    v["tracks"]["items"]
-        .as_array()
-        .into_iter()
-        .flatten()
-        // Search matches loosely — keep only tracks this artist is credited on.
-        .filter(|t| {
-            t["artists"]
-                .as_array()
-                .is_some_and(|a| a.iter().any(|x| x["id"].as_str() == Some(id)))
-        })
-        .filter_map(|t| {
-            Some(LibItem::track(
-                t["name"].as_str()?.to_string(),
-                t["artists"][0]["name"].as_str().unwrap_or("").to_string(),
-                t["uri"].as_str()?.to_string(),
-            ))
-        })
-        .collect()
-}
-
-/// An artist's albums and singles, deduped by name, newest first. Paged ten at
-/// a time — since February 2026 a larger page is a 400 "Invalid limit".
-fn artist_albums(client: &reqwest::blocking::Client, token: &str, id: &str) -> Vec<LibItem> {
-    let mut seen = std::collections::HashSet::new();
-    let mut albums: Vec<(String, LibItem)> = Vec::new();
-    let mut url = Some(format!(
-        "{API}/artists/{id}/albums?include_groups=album,single&limit=10"
-    ));
-    let mut pages = 0;
-    while let Some(u) = url.take() {
-        if pages >= 3 {
-            break; // 30 releases; more pages burn the API quota per drill-in
-        }
-        let Some(v) = get_json_cached(client, &u, token) else {
-            break;
-        };
-        for a in v["items"].as_array().into_iter().flatten() {
-            let (Some(aname), Some(auri)) = (a["name"].as_str(), a["uri"].as_str()) else {
-                continue;
-            };
-            if !seen.insert(aname.to_lowercase()) {
-                continue;
-            }
-            let date = a["release_date"].as_str().unwrap_or("").to_string();
-            let year = date.split('-').next().unwrap_or("").to_string();
-            albums.push((
-                date,
-                LibItem::ctx(aname.to_string(), year, auri.to_string()),
-            ));
-        }
-        url = v["next"].as_str().map(String::from);
-        pages += 1;
-    }
-    albums.sort_by(|x, y| y.0.cmp(&x.0)); // newest first
-    albums.into_iter().map(|(_, it)| it).collect()
-}
-
-fn fetch_detail_blocking(token: &str, uri: &str, name: &str) -> (String, Vec<LibItem>) {
-    let client = http_client();
-    let mut parts = uri.split(':');
-    parts.next(); // "spotify"
-    let kind = parts.next().unwrap_or("");
-    let id = parts.next().unwrap_or("");
-
-    // "Play all" row first.
-    let mut items = vec![LibItem::play(format!("▶︎ Play {name}"), uri.to_string())];
-
-    liblog(format!("detail: kind={kind} id={id} uri={uri}"));
-    match kind {
-        "artist" => {
-            let tracks = artist_top_tracks(&client, token, id, name);
-            if !tracks.is_empty() {
-                items.push(LibItem::header("Popular"));
-                items.extend(tracks);
-            }
-            let albums = artist_albums(&client, token, id);
-            if !albums.is_empty() {
-                items.push(LibItem::header("Albums"));
-                items.extend(albums);
-            }
-            if items.len() == 1 {
-                // Only the play row: every request failed, most often a spent
-                // API quota. Say so instead of showing a bare button.
-                items.push(LibItem::header("nothing loaded — API error or quota"));
-            }
-        }
-        "album" => {
-            if let Some(v) = get_json_cached(
-                &client,
-                &format!("{API}/albums/{id}/tracks?limit=50"),
-                token,
-            ) {
-                for t in v["items"].as_array().into_iter().flatten() {
-                    if let (Some(n), Some(u)) = (t["name"].as_str(), t["uri"].as_str()) {
-                        items.push(LibItem::track(
-                            n.to_string(),
-                            t["artists"][0]["name"].as_str().unwrap_or("").to_string(),
-                            u.to_string(),
-                        ));
-                    }
-                }
-            }
-        }
-        "playlist" => {
-            // Follow `next` instead of taking only the first page: playlists
-            // routinely exceed the 100-item page size, and the drill-in list
-            // (plus "play from this track") was silently truncated.
-            let before = items.len();
-            items.extend(fetch_all_pages(
-                &client,
-                &format!("{API}/playlists/{id}/items?limit=100"),
-                token,
-                None, // items[] is top-level on this endpoint
-                10,   // 1,000 tracks, matching the other sections' ceiling
-                parse_playlist_track,
-            ));
-            if items.len() == before {
-                // Some third-party playlists 403 even on /items. `fetch_all_pages`
-                // has no error channel, so an empty result is indistinguishable
-                // from an empty playlist — say both rather than showing a blank
-                // pane with no explanation.
-                items.push(LibItem::header("no tracks — empty or restricted"));
-            }
-        }
-        _ => {}
-    }
-
-    liblog(format!("detail: {} rows for {uri}", items.len()));
-    (name.to_string(), items)
-}
-
-// --- Live playback state (server-side) ---
-
-/// The current playback as Spotify remembers it (across devices).
-struct RemotePlaybackState {
-    track_id: String,
-    progress_ms: u32,
-    shuffle: bool,
-    repeat: bool,
-    volume: u8,
-}
-
-fn fetch_playback_state(token: &str) -> Option<RemotePlaybackState> {
-    let client = http_client();
-    let resp = client
-        .get(format!("{API}/me/player"))
-        .bearer_auth(token)
-        .send()
-        .ok()?;
-    if !resp.status().is_success() {
-        return None; // 204 = nothing playing recently
-    }
-    let v: serde_json::Value = resp.json().ok()?;
-    let track_id = v["item"]["id"].as_str()?.to_string();
-    Some(RemotePlaybackState {
-        track_id,
-        progress_ms: v["progress_ms"].as_u64().unwrap_or(0) as u32,
-        shuffle: v["shuffle_state"].as_bool().unwrap_or(false),
-        repeat: v["repeat_state"]
-            .as_str()
-            .map(|r| r != "off")
-            .unwrap_or(false),
-        volume: v["device"]["volume_percent"].as_u64().unwrap_or(50) as u8,
-    })
-}
-
-/// Transfer the current server-side playback onto the myx device (with its full
-/// context + queue + position). `play=false` transfers paused.
-/// `Ok(())` on success, `Err(reason)` with something worth showing otherwise.
-fn transfer_playback(token: &str, device_id: &str, play: bool) -> Result<(), String> {
-    let client = http_client();
-    match client
-        .put(format!("{API}/me/player"))
-        .bearer_auth(token)
-        .json(&serde_json::json!({ "device_ids": [device_id], "play": play }))
-        .send()
-    {
-        Ok(r) if r.status().is_success() => Ok(()),
-        Ok(r) => {
-            let code = r.status().as_u16();
-            let body = r.text().unwrap_or_default();
-            liblog(format!("transfer -> HTTP {code}: {body}"));
-            Err(format!("HTTP {code}"))
-        }
-        Err(e) => {
-            liblog(format!("transfer failed: {e}"));
-            Err("network error".to_string())
-        }
-    }
-}
-
-/// Boot restore: read the live playback state, transfer it onto myx (retrying
-/// while the device registers), and hand the state back to the UI.
-fn spawn_restore(
-    webapi: Arc<Mutex<WebApi>>,
-    device_id: String,
-    tx: flume::Sender<RemotePlaybackState>,
-) {
-    tokio::task::spawn_blocking(move || {
-        let Some(token) = token_of(&webapi) else {
-            return;
-        };
-        let Some(state) = fetch_playback_state(&token) else {
-            return;
-        };
-        // Retry the transfer — the Connect device can take a moment to appear.
-        for _ in 0..6 {
-            if transfer_playback(&token, &device_id, false).is_ok() {
-                break;
-            }
-            std::thread::sleep(Duration::from_secs(1));
-        }
-        let _ = tx.send(state);
-    });
-}
-
-// --- Live playback state (server-side) end ---
-
-/// A blocking HTTP client with a timeout so a stalled network can't wedge a
-/// worker thread forever (audit H2).
-fn http_client() -> reqwest::blocking::Client {
-    reqwest::blocking::Client::builder()
-        .timeout(Duration::from_secs(10))
-        .build()
-        .unwrap_or_default()
 }
 
 // ------------------------------------------------------------------ terminal
